refactor: split components.js into src/components/
components.js had reached 1522 lines and 13 exported components. Split one
module per widget, with the shared plumbing separated out:

  theme.js      palette, ThemeProvider, useTheme, useControl, labelContent
  anchor.js     anchorRect/useAnchor, screenOf, measureLabel
  typeahead.js  useTypeAhead, typeAheadChar
  keys.js       X11 keysym constants
  index.js      re-exports the public names

Largest module is now Menu.js at 547 lines; most are under 120.

Move only — no behaviour change. The split was done mechanically by
slicing the original file per top-level symbol rather than by hand, and
verified two ways:

- the public surface is byte-identical before and after (named exports of
  src/index.js, its default export's keys, and the components barrel with
  each export's typeof);
- diffing every non-import, non-comment line of the old file against the
  concatenated new modules leaves zero real differences — the only added
  lines are the per-module `const h = React.createElement` alias and the
  import statements.

src/components.js was not a package entry point (package.json exports only
"."), so no compatibility shim is needed.
