+++ src/components/index.js
@@ -1,0 +1,15 @@
+// Public widget components. Split out of a single components.js; the
+// modules here are grouped by widget, with the shared plumbing in
+// theme.js (palette + control props), anchor.js (popup geometry),
+// typeahead.js and keys.js.
+export { ThemeProvider, SelectThemeProvider } from './theme.js';
+export { anchorRect, useAnchor } from './anchor.js';
+export { Button } from './Button.js';
+export { Checkbox } from './Checkbox.js';
+export { Radio, RadioGroup } from './Radio.js';
+export { Switch } from './Switch.js';
+export { ProgressBar } from './ProgressBar.js';
+export { Slider } from './Slider.js';
+export { Tooltip } from './Tooltip.js';
+export { ContextMenu, MenuBar } from './Menu.js';
+export { Select } from './Select.js';
