+++ src/components/keys.js
@@ -1,0 +1,21 @@
+// X11 keysyms used by the widget keyboard handlers.
+
+export const XK_RETURN = 0xff0d;
+
+export const XK_ESCAPE = 0xff1b;
+
+export const XK_HOME = 0xff50;
+
+export const XK_LEFT = 0xff51;
+
+export const XK_UP = 0xff52;
+
+export const XK_RIGHT = 0xff53;
+
+export const XK_DOWN = 0xff54;
+
+export const XK_PAGE_UP = 0xff55;
+
+export const XK_PAGE_DOWN = 0xff56;
+
+export const XK_END = 0xff57;
