+++ src/index.js
@@ -20,7 +20,7 @@ export {
   ContextMenu,
   useAnchor,
   anchorRect,
-} from './components.js';
+} from './components/index.js';
 
 import { render, createRoot, unmountComponentAtNode } from './Reconciler.js';
